+++ man-roxygen/roxlate-ml-decision-trees-min-info-gain.R
@@ -1,0 +1,1 @@
+#' @param min.info.gain Minimum information gain for a split to be considered at a tree node. Should be >= 0, defaults to 0.